backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv
pymongo==4.8.0